update the exception status in user and todo controller

+++ src/todo/todo.service.ts
@@ -1,4 +1,10 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common';
+import {
+  BadRequestException,
+  HttpException,
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
 import { Todo } from 'src/entities/todo.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
@@ -10,19 +16,27 @@ export class TodoService {
   ) {}
 
   // get all todo entries from database
-  async getTodos(): Promise<Todo[]> {
+  async getTodos(): Promise<Todo[] | HttpException> {
     try {
-      const Todo = await this.todoRepository.find();
-      return Todo;
+      const Todos = await this.todoRepository.find();
+      // return not found exception if database is empty
+      if (!Todos.length) {
+        return new NotFoundException('No Todos in database exist');
+      }
+      return Todos;
     } catch (error) {
       throw new InternalServerErrorException();
     }
   }
 
   // get a todo entry by ID from database
-  async getTodo(todoId: number): Promise<Todo> {
+  async getTodo(todoId: number): Promise<Todo | HttpException> {
     try {
       const todo = await this.todoRepository.findOne({ where: { id: todoId } });
+      // return not found exception if todo with id does not exists
+      if (!todo) {
+        return new NotFoundException(`No todo with id:${todoId} exists`);
+      }
       return todo;
     } catch (error) {
       throw new InternalServerErrorException();
@@ -44,6 +58,14 @@ export class TodoService {
   // delete a todo entry from database
   async deleteTodo(id: number): Promise<any> {
     try {
+      // check if todo with given id exists
+      const todo = await this.todoRepository.findOne({ where: { id } });
+
+      // return error if todo with id does not exists
+      if (!todo) {
+        return new BadRequestException({ description: 'User does not exist' });
+      }
+      // delete the todo
       await this.todoRepository.delete({ id });
       return { description: `Todo deleted successfully` };
     } catch (error) {
